pyproject.toml
[project]
name = "content"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "apify-client>=1.8.0",
    "fastapi>=0.135.1",
    "google-cloud-videointelligence>=2.13.0",
    "google-genai>=1.5.0",
    "httpx>=0.28.1",
    "openai>=1.60.0",
    "pillow>=12.1.1",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.48",
    "uvicorn[standard]>=0.41.0",
    "websockets>=16.0",
    "yt-dlp>=2026.3.13",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.5",
]

# --- Ruff ---
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "C4", "SIM", "TCH"]

[tool.ruff.lint.isort]
known-first-party = ["backend"]

# --- Mypy ---
[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["aiofiles.*", "aiosqlite.*", "apify_client.*", "google.*", "openai.*", "yt_dlp.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

# --- Pytest ---
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: end-to-end tests requiring live services (Gemini, ComfyUI). Run with: uv run pytest -m integration",
]
addopts = "-m 'not integration'"
filterwarnings = [
    "ignore::DeprecationWarning:sqlalchemy.*",
]
